--- src/main/java/cs414f20/teamd/PawnTest.java
@@ -1,32 +1,0 @@
-package a2;
-
-import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.Test;
-
-import a2.ChessPiece.Color;
-
-class PawnTest {
-
-	@Test
-	void testDefaultPawnColor() {
-		ChessBoard testBoard = new ChessBoard();
-		Pawn testPawn = new Pawn();
-		Pawn testBlackPawn = new Pawn(testBoard, Color.BLACK);
-		
-		assertTrue(testPawn.color == Color.WHITE);
-		assertTrue(testBlackPawn.color == Color.BLACK);
-	}
-	
-	@Test
-	void testDefaultPawnString() {
-		ChessBoard testBoard = new ChessBoard();
-		Pawn testPawn = new Pawn();
-		Pawn testBlackPawn = new Pawn(testBoard, Color.BLACK);
-		String testWhitePawnString = testPawn.toString();
-		String testBlackPawnString = testBlackPawn.toString();
-		
-		assertTrue(testWhitePawnString == "\u2659");
-		assertTrue(testBlackPawnString == "\u265F");
-	}
-	
-}
