--- src/main/java/cs414f20/teamd/RookTest.java
@@ -1,32 +1,0 @@
-package cs414f20.teamd;
-
-import static org.junit.jupiter.api.Assertions.*;
-import org.junit.jupiter.api.Test;
-
-import a2.ChessPiece.Color;
-
-class RookTest {
-
-	@Test
-	void testDefaultRookColor() {
-		ChessBoard testBoard = new ChessBoard();
-		Rook testRook = new Rook();
-		Rook testBlackRook = new Rook(testBoard, Color.BLACK);
-		
-		assertTrue(testRook.color == Color.WHITE);
-		assertTrue(testBlackRook.color == Color.BLACK);
-	}
-	
-	@Test
-	void testDefaultRookString() {
-		ChessBoard testBoard = new ChessBoard();
-		Rook testRook = new Rook();
-		Rook testBlackRook = new Rook(testBoard, Color.BLACK);
-		String testWhiteRookString = testRook.toString();
-		String testBlackRookString = testBlackRook.toString();
-		
-		assertTrue(testWhiteRookString == "\u2656");
-		assertTrue(testBlackRookString == "\u265C");
-	}
-
-}
