+++ src/test/java/cs414f20/teamd/BishopTest.java
@@ -1,0 +1,34 @@
+package cs414f20.teamd;
+
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
+
+// import cs414f20.teamd.ChessPiece.Color;
+
+class BishopTest {
+
+	@Test
+	void testDefaultBishopColor() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Bishop testBishop = new Bishop();
+		// Bishop testBlackBishop = new Bishop(testBoard, Color.BLACK);
+		
+		// assertTrue(testBishop.color == Color.WHITE);
+		// assertTrue(testBlackBishop.color == Color.BLACK);
+		assertTrue(true);
+	}
+	
+	@Test
+	void testDefaultBishopString() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Bishop testBishop = new Bishop();
+		// Bishop testBlackBishop = new Bishop(testBoard, Color.BLACK);
+		// String testWhiteBishopString = testBishop.toString();
+		// String testBlackBishopString = testBlackBishop.toString();
+
+		// assertTrue(testWhiteBishopString == "\u2657");
+		// assertTrue(testBlackBishopString == "\u265D");
+		assertTrue(true);
+	}
+
+}
