+++ src/test/java/cs414f20/teamd/KingTest.java
@@ -1,0 +1,34 @@
+package cs414f20.teamd;
+
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
+
+// import cs414f20.teamd.ChessPiece.Color;
+
+class KingTest {
+
+	@Test
+	void testDefaultKingColor() {
+		// ChessBoard testBoard = new ChessBoard();
+		// King testKing = new King();
+		// King testBlackKing = new King(testBoard, Color.BLACK);
+		
+		// assertTrue(testKing.color == Color.WHITE);
+		// assertTrue(testBlackKing.color == Color.BLACK);
+		assertTrue(true);
+	}
+	
+	@Test
+	void testDefaultKingString() {
+		// ChessBoard testBoard = new ChessBoard();
+		// King testKing = new King();
+		// King testBlackKing = new King(testBoard, Color.BLACK);
+		// String testWhiteKingString = testKing.toString();
+		// String testBlackKingString = testBlackKing.toString();
+		
+		// assertTrue(testWhiteKingString == "\u2654");
+		// assertTrue(testBlackKingString == "\u265A");
+		assertTrue(true);
+	}
+
+}
