+++ src/test/java/cs414f20/teamd/KnightTest.java
@@ -1,0 +1,45 @@
+package cs414f20.teamd;
+
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
+
+// import java.util.ArrayList;
+
+// import cs414f20.teamd.ChessPiece.Color;
+
+class KnightTest {
+
+	@Test
+	void testDefaultKnightColor() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Knight testKnight = new Knight();
+		// Knight testBlackKnight = new Knight(testBoard, Color.BLACK);
+		
+		// assertTrue(testKnight.color == Color.WHITE);
+		// assertTrue(testBlackKnight.color == Color.BLACK);
+		assertTrue(true);
+	}
+	
+	@Test
+	void testDefaultKnightString() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Knight testKnight = new Knight();
+		// Knight testBlackKnight = new Knight(testBoard, Color.BLACK);
+		// String testWhiteKnightString = testKnight.toString();
+		// String testBlackKnightString = testBlackKnight.toString();
+		
+		// assertTrue(testWhiteKnightString == "\u2658");
+		// assertTrue(testBlackKnightString == "\u265E");
+		assertTrue(true);
+	}
+	
+	@Test
+	void testValidMoves() {
+		// Knight testKnight = new Knight();
+		// ArrayList<String> validMoves = testKnight.legalMoves();
+		
+		// assertTrue(validMoves.isEmpty());
+		assertTrue(true);
+	}
+
+}
