+++ src/test/java/cs414f20/teamd/QueenTest.java
@@ -1,0 +1,44 @@
+package cs414f20.teamd;
+
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
+// import java.util.ArrayList;
+
+// import cs414f20.teamd.ChessPiece.Color;
+
+class QueenTest {
+
+	@Test
+	void testDefaultQueenColor() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Queen testQueen = new Queen();
+		// Queen testBlackQueen = new Queen(testBoard, Color.BLACK);
+		
+		// assertTrue(testQueen.color == Color.WHITE);
+		// assertTrue(testBlackQueen.color == Color.BLACK);
+		assertTrue(true);
+	}
+	
+	@Test
+	void testDefaultQueenString() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Queen testQueen = new Queen();
+		// Queen testBlackQueen = new Queen(testBoard, Color.BLACK);
+		// String testWhiteQueenString = testQueen.toString();
+		// String testBlackQueenString = testBlackQueen.toString();
+		
+		// assertTrue(testWhiteQueenString == "\u2655");
+		// assertTrue(testBlackQueenString == "\u265B");
+		assertTrue(true);
+	}
+	
+	@Test
+	void testLegalMoves() {
+		// Queen testQueen = new Queen();
+		// ArrayList<String> validMoves = testQueen.legalMoves();
+		
+		// assertTrue(validMoves.isEmpty());
+		assertTrue(true);
+	}
+
+}
