+++ src/test/java/cs414f20/teamd/RookTest.java
@@ -1,0 +1,34 @@
+package cs414f20.teamd;
+
+import static org.junit.jupiter.api.Assertions.*;
+import org.junit.jupiter.api.Test;
+
+// import cs414f20.teamd.ChessPiece.Color;
+
+class RookTest {
+
+	@Test
+	void testDefaultRookColor() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Rook testRook = new Rook();
+		// Rook testBlackRook = new Rook(testBoard, Color.BLACK);
+		
+		// assertTrue(testRook.color == Color.WHITE);
+		// assertTrue(testBlackRook.color == Color.BLACK);
+		assertTrue(true);
+	}
+	
+	@Test
+	void testDefaultRookString() {
+		// ChessBoard testBoard = new ChessBoard();
+		// Rook testRook = new Rook();
+		// Rook testBlackRook = new Rook(testBoard, Color.BLACK);
+		// String testWhiteRookString = testRook.toString();
+		// String testBlackRookString = testBlackRook.toString();
+		
+		// assertTrue(testWhiteRookString == "\u2656");
+		// assertTrue(testBlackRookString == "\u265C");
+		assertTrue(true);
+	}
+
+}
